requirements.txt
djangorestframework~=3.12.4
Django~=3.2.2
psycopg2-binary~=2.8.6